requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==0.5.2
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
av==12.3.0
beautifulsoup4==4.12.3
bidict==0.23.1
black==24.4.2
blinker==1.8.2
braceexpand==0.1.7
bs4==0.0.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
coloredlogs==15.0.1
concurrent-log-handler==0.9.25
contourpy==1.2.1
controlnet-aux==0.0.9
cycler==0.12.1
Cython==3.0.11
diffusers @ git+https://github.com/huggingface/diffusers.git@11542431a52b02ef4f14b6f53354c79187884827
dill==0.3.9
docker-pycreds==0.4.0
dominate==2.9.1
easydict==1.9
einops==0.8.0
exceptiongroup==1.2.2
fastapi==0.112.0
fastrlock==0.8.2
ffmpy==0.4.0
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
fonttools==4.53.1
freetype-py==2.5.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.0
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.4.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.33.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==4.40.0
gradio_client==1.2.0
graphviz==0.8.4
grpcio==1.65.1
h11==0.14.0
h5py==3.11.0
hezar==0.40.3
hf-cli==0.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
humanfriendly==10.0
hydra-core==1.3.2
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
imgaug==0.4.0
importlib_metadata==8.2.0
importlib_resources==6.4.0
invisible-watermark==0.2.0
iopath==0.1.9
ip-adapter @ git+https://github.com/tencent-ailab/IP-Adapter.git@62e4af9d0c1ac7d5f8dd386a0ccf2211346af1a2
itsdangerous==2.2.0
jax==0.4.31
jaxlib==0.4.31
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.5.0
kiwisolver==1.4.5
kornia==0.5.0
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.11.6
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mediapipe==0.10.15
ml-dtypes==0.3.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
open3d-python==0.3.0.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
peft==0.12.0
Pillow==10.0.0
platformdirs==4.2.2
portalocker==2.10.1
prettytable==3.11.0
proglog==0.1.10
protobuf==4.25.4
psutil==6.0.0
pyamg==5.2.1
pyarrow==12.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
pyglet==2.0.17
Pygments==2.18.0
PyOpenGL==3.1.0
pyparsing==3.1.2
pyrender==0.1.45
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
python-telegram-bot==21.6
python-version==0.0.2
pytorch-lightning==1.2.9
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
ruff==0.5.6
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.15.0
setproctitle==1.3.3
shapely==2.0.5
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
smplx==0.1.28
sniffio==1.3.1
sounddevice==0.5.0
soupsieve==2.5
starlette==0.37.2
streamlit==1.37.1
streamlit-keyup==0.2.4
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
#tensorflow==2.17.0
#tensorflow-estimator==2.14.0
#tensorflow-io-gcs-filesystem==0.37.1
tensorrt==10.3.0
tensorrt_dispatch==10.3.0
tensorrt_lean==10.3.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.7.24
timm==0.6.7
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torchaudio==2.4.0
torchmetrics==0.2.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
transformers==4.42.4
trimesh==4.4.9
triton==3.0.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.5
wandb==0.18.3
watchdog==4.0.1
wcwidth==0.2.13
webdataset==0.2.86
websockets==12.0
Werkzeug==3.0.3
wldhx.yadisk-direct==0.0.6
wrapt==1.14.1
wsproto==1.2.0
xformers==0.0.27.post2
yacs==0.1.8
yarl==1.9.4
zipp==3.19.2
